Added type safety

+++ Engine/GameEngine.cpp
@@ -34,7 +34,11 @@ void GameEngine::SetMap(std::unique_ptr<GameMap> game_map)
 	m_GameMap = std::move(game_map);
 	if (m_GameMap)
 	{
-		m_GameMap->SetSceneBounds(m_WindowWidth, m_WindowHeight);
+		m_GameMap->SetSceneBounds
+		(
+			static_cast<float>(m_WindowWidth), 
+			static_cast<float>(m_WindowHeight)
+		);
 		m_GameMap->Initialize();
 	}
 }
